+++ src/utilities/piral-debug-utils/src/debug.ts
@@ -143,8 +143,16 @@ export function installPiralDebug(options: DebuggerOptions) {
   const togglePilet = (name: string) => {
     const pilet: any = getPilets().find((m) => m.name === name);
 
-    if (pilet.disabled) {
-      updatePilet(pilet.original);
+    if (!pilet) {
+      // nothing to do, obviously invalid call
+    } else if (pilet.disabled) {
+      if (pilet.original) {
+        // everything is fine, let's use the cached version
+        updatePilet(pilet.original);
+      } else {
+        // something fishy is going on - let's just try to activate the same pilet
+        updatePilet({ ...pilet, disabled: false });
+      }
     } else {
       updatePilet({ name, disabled: true, original: pilet });
     }
@@ -183,17 +191,19 @@ export function installPiralDebug(options: DebuggerOptions) {
   const eventDispatcher = document.body.dispatchEvent;
 
   const systemResolve = System.constructor.prototype.resolve;
-  const depMap = {};
+  const depMap: Record<string, Record<string, string>> = {};
 
   System.constructor.prototype.resolve = function (...args) {
     const [url, parent] = args;
+    const result = systemResolve.call(this, ...args);
 
     if (parent) {
-      depMap[parent] = depMap[parent] || {};
-      depMap[parent][url] = true;
+      const deps = depMap[parent] || {};
+      deps[url] = result;
+      depMap[parent] = deps;
     }
 
-    return systemResolve.call(this, ...args);
+    return result;
   };
 
   const debugApi = {
@@ -262,13 +272,20 @@ export function installPiralDebug(options: DebuggerOptions) {
   };
 
   const getDependencyMap = () => {
-    const dependencyMap: Record<string, Array<string>> = {};
-    const addDeps = (pilet: string, dependencies: Array<string>) => {
-      if (!(pilet in dependencyMap)) {
-        dependencyMap[pilet] = [];
+    const dependencyMap: Record<string, Array<{ demanded: string; resolved: string }>> = {};
+    const addDeps = (pilet: string, dependencies: Record<string, string>) => {
+      const deps = dependencyMap[pilet] || [];
+
+      for (const depName of Object.keys(dependencies)) {
+        if (!deps.some((m) => m.demanded === depName)) {
+          deps.push({
+            demanded: depName,
+            resolved: dependencies[depName],
+          });
+        }
       }
 
-      dependencyMap[pilet].push(...dependencies);
+      dependencyMap[pilet] = deps;
     };
     const pilets = getPilets()
       .map((pilet: any) => ({