+++ CHANGELOG.md
@@ -18,6 +18,12 @@
 - Added types to `piral-cli-webpack/extend-config` and `piral-cli-webpack5/extend-config`
 - Added semver possibilities for resolving shared dependencies
 
+## 0.14.15 (March 5, 2022)
+
+- Fixed issue with `get-dependency-map` of the `piral-debug-utils`
+- Added support for v0 and v1 pilets in `get-dependency-map`
+- Added support for object flags (e.g., `--vars`) in the initializers
+
 ## 0.14.14 (March 3, 2022)
 
 - Fixed issue placing *tsconfig.json* in empty template